Created LICENSE.md
Initial commit of a self made license file with the intention of keeping this repo public accessible and usable by all of humanity not just the few. 

This license file is intended to serve as a first line of defense against bad actors who would seek to use this knowledge for their own gains whilst ignoring the plight of others.